+++ src/api/java/reborncore/api/power/IPowerConfig.java
@@ -1,0 +1,13 @@
+package reborncore.api.power;
+
+public interface IPowerConfig {
+
+	boolean eu();
+
+	boolean tesla();
+
+	boolean internal();
+
+	boolean forge();
+
+}