+++ src/main/java/szewek/mcflux/wrapper/EnergyType.java
@@ -3,7 +3,7 @@ package szewek.mcflux.wrapper;
 import szewek.mcflux.util.MCFluxLocation;
 
 public enum EnergyType {
-	FORGE_ENERGY("fe"), TESLA("tesla"), EU("eu"), IF("if"), RF("rf"), NONE("");
+	FORGE_ENERGY("fe"), TESLA("tesla"), MEKANISM("mkj"), REBORN_CORE("rc"), EU("eu"), IF("if"), RF("rf"), NONE("");
 
 	public static final EnergyType[] ALL;